Create a newHook for fadestyles
