+++ src/components/SubNav/SubNav.test.tsx
@@ -1,6 +1,5 @@
 import { render, screen } from "@testing-library/react";
 import { axe } from "jest-axe";
-import Icon from "../Icons/Icon";
 import SubNav, { SubNavButton, SubNavLink } from "./SubNav";
 
 describe("SubNav Accessibility", () => {
@@ -20,69 +19,8 @@ describe("SubNav Accessibility", () => {
             </SubNavButton>
           </>
         }
-        secondaryActions={
-          <>
-            <SubNavLink
-              id="primary-sub-nav-link-1"
-              screenreaderOnlyText="for the NYPL Research Catalog"
-              href="#link1"
-            >
-              Label Text
-            </SubNavLink>
-            <SubNavLink
-              id="primary-sub-nav-link-2"
-              screenreaderOnlyText="for the NYPL Research Catalog"
-              href="#link2"
-            >
-              Label Text
-            </SubNavLink>
-            <SubNavLink
-              id="primary-sub-nav-link-3"
-              screenreaderOnlyText="for the NYPL Research Catalog"
-              href="#link3"
-            >
-              Label Text
-              <Icon
-                name="legacyAccountFilled"
-                size="small"
-                className="right"
-                align="right"
-              />
-            </SubNavLink>
-          </>
-        }
       />
     );
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it("passes axe accessibility test with secondary actions", async () => {
-    const { container } = render(
-      <SubNav
-        primaryActions={
-          <>
-            <SubNavButton id="primary-sub-nav-button-1" buttonType="text">
-              Label Text
-            </SubNavButton>
-          </>
-        }
-        secondaryActions={
-          <>
-            <SubNavButton id="secondary-sub-nav-button-1" buttonType="text">
-              Label Text
-            </SubNavButton>
-            <SubNavLink
-              id="secondary-sub-nav-link-1"
-              screenreaderOnlyText="for the NYPL Research Catalog"
-              href="#link1"
-            >
-              Label Text
-            </SubNavLink>
-          </>
-        }
-      />
-    );
-
     expect(await axe(container)).toHaveNoViolations();
   });
 
@@ -172,76 +110,6 @@ describe("SubNav Component", () => {
 
     expect(screen.getByText("Primary Button")).toBeInTheDocument();
     expect(screen.getByText("Secondary Link")).toBeInTheDocument();
-  });
-
-  it("renders with selectedItem highlighted", async () => {
-    render(
-      <SubNav
-        primaryActions={
-          <>
-            <SubNavButton id="primary-button-1" isSelected>
-              Primary Button 1
-            </SubNavButton>
-            <SubNavButton id="primary-button-2">Primary Button 2</SubNavButton>
-          </>
-        }
-      />
-    );
-
-    const selectedButton = await screen.findByRole("button", {
-      name: /Primary Button 1/i,
-    });
-    expect(selectedButton).toHaveTextContent("Primary Button 1");
-  });
-
-  it("passes axe accessibility test with primary actions", async () => {
-    const { container } = render(
-      <SubNav
-        primaryActions={
-          <>
-            <SubNavButton id="primary-button-1">Primary Button 1</SubNavButton>
-            <SubNavButton id="primary-button-2">Primary Button 2</SubNavButton>
-          </>
-        }
-      />
-    );
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it("passes axe accessibility test with secondary actions", async () => {
-    const { container } = render(
-      <SubNav
-        primaryActions={
-          <>
-            <SubNavButton id="primary-button-1">Primary Button</SubNavButton>
-          </>
-        }
-        secondaryActions={
-          <>
-            <SubNavLink href="#secondary-link">Secondary Link</SubNavLink>
-          </>
-        }
-      />
-    );
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it("passes axe accessibility test with props", async () => {
-    const { container } = render(
-      <SubNav
-        actionBackgroundColor="brand.primary-05"
-        highlightColor="brand.primary"
-        id="sub-nav-id"
-        primaryActions={
-          <>
-            <SubNavButton id="primary-sub-nav-button-1">
-              Label Text
-            </SubNavButton>
-          </>
-        }
-      />
-    );
-    expect(await axe(container)).toHaveNoViolations();
   });
 });
